Refactor: Reorganizando modulos da aplicacao.
