+++ myfinance/purchase/tests/tests.py
@@ -1,5 +1,5 @@
 from django.test import TestCase, Client
-from purchase import views, models
+from purchase import views, models, utils
 import mock
 from StringIO import StringIO
 
@@ -10,13 +10,13 @@ class PurchaseImporterFormViewTestCase(TestCase):
     def setUpClass(cls):
         cls.client = Client()
 
-    @mock.patch.object(views, 'import_data')
+    @mock.patch.object(utils, 'import_data')
     def test_if_purchase_importer_form_calls_import_data_in_case_of_post_request(self, import_data_mock):
         self.client.post('/purchase/pdi/')
         self.assertTrue(import_data_mock.called)
 
     @mock.patch('purchase.views.render')
-    @mock.patch.object(views, 'import_data')
+    @mock.patch.object(utils, 'import_data')
     def test_purchase_importer_form_must_render_the_template_with_import_feedback_and_the_purchase_total_messages_in_case_of_post_request(self,
             import_data_mock, render_mock):
         fake_import_data_messages = {'import_feedback': 'Foo', 'purchase_total': 'Bar'}
@@ -31,7 +31,7 @@ class PurchaseImporterFormViewTestCase(TestCase):
             fake_import_data_messages,
         )
 
-    @mock.patch.object(views, 'import_data')
+    @mock.patch.object(utils, 'import_data')
     def test_purchase_importer_form_should_not_calls_import_data_in_case_of_get_request(self, import_data_mock):
         self.client.get('/purchase/pdi/')
         self.assertFalse(import_data_mock.called)
@@ -39,10 +39,10 @@ class PurchaseImporterFormViewTestCase(TestCase):
 
 class ImportDataViewTestCase(TestCase):
 
-    @mock.patch.object(views, 'calc_purchase_total')
-    @mock.patch.object(views, 'save_purchase_data')
-    @mock.patch.object(views, 'read_file_lines')
-    @mock.patch.object(views, 'parse_purchase_file_data')
+    @mock.patch.object(utils, 'calc_purchase_total')
+    @mock.patch.object(utils, 'save_purchase_data')
+    @mock.patch.object(utils, 'read_file_lines')
+    @mock.patch.object(utils, 'parse_purchase_file_data')
     @mock.patch.object(models, 'Purchase')
     def test_import_data_should_return_a_save_feedback_message_and_the_purchase_total(self,
             purchase_model_mock, parse_purchase_file_data_mock, read_file_lines_mock,
@@ -57,7 +57,7 @@ class ImportDataViewTestCase(TestCase):
 
         calc_purchase_total_mock.return_value = 20
 
-        messages = views.import_data(request_mock)
+        messages = utils.import_data(request_mock)
 
         self.assertIn('import_feedback', messages)
         self.assertIn('purchase_total', messages)
@@ -65,10 +65,10 @@ class ImportDataViewTestCase(TestCase):
         self.assertEquals(messages['import_feedback'], 'Importacao efetuada com sucesso.')
         self.assertEquals(messages['purchase_total'], 'A receita bruta total foi de R$ 20.')
 
-    @mock.patch.object(views, 'calc_purchase_total')
-    @mock.patch.object(views, 'save_purchase_data')
-    @mock.patch.object(views, 'read_file_lines')
-    @mock.patch.object(views, 'parse_purchase_file_data')
+    @mock.patch.object(utils, 'calc_purchase_total')
+    @mock.patch.object(utils, 'save_purchase_data')
+    @mock.patch.object(utils, 'read_file_lines')
+    @mock.patch.object(utils, 'parse_purchase_file_data')
     def test_import_data_integration(self, parse_purchase_file_data_mock,
             read_file_lines_mock, save_purchase_data_mock, calc_purchase_total_mock):
         request_mock = mock.Mock()
@@ -88,7 +88,7 @@ class ImportDataViewTestCase(TestCase):
         }
         parse_purchase_file_data_mock.return_value = parsed_data
 
-        views.import_data(request_mock)
+        utils.import_data(request_mock)
 
         read_file_lines_mock.assert_called_with(fake_purchase_file)
 
@@ -101,7 +101,7 @@ class ImportDataViewTestCase(TestCase):
 
 class ParsePurchaseFileDataViewTestCase(TestCase):
 
-    @mock.patch.object(views, 'parse_purchase_file_data')
+    @mock.patch.object(utils, 'parse_purchase_file_data')
     def test_parse_purchase_file_data_should_transform_each_line_in_a_tuple(self,
             parse_purchase_file_data_mock):
         fake_purchase_file = StringIO()
@@ -109,7 +109,7 @@ class ParsePurchaseFileDataViewTestCase(TestCase):
         merchant   address merchant name
         Joao Silva  R$10 off R$20 of food   10.0    2   987 Fake St Bobs Pizza""")
         expected_parsed_line = ('Joao Silva', 'R$10 off', 'R$ 20 of food', '10.0', '2', '987 Fake St', 'Bobs Pizza')
-        for parsed_line in views.parse_purchase_file_data(fake_purchase_file):
+        for parsed_line in utils.parse_purchase_file_data(fake_purchase_file):
             self.assertEquals(parsed_line, expected_parsed_line)
         self.assertTrue(parse_purchase_file_data_mock.called)
 
@@ -141,7 +141,7 @@ class SavePurchaseDataViewTestCase(TestCase):
             'merchant_address': '987 Fake St',
             'merchant_name': 'Bobs Pizza'
         }
-        views.save_purchase_data(parsed_data)
+        utils.save_purchase_data(parsed_data)
 
         purchaser_create_mock.assert_called_with(name=parsed_data['purchaser_name'])
 
@@ -158,4 +158,4 @@ class SavePurchaseDataViewTestCase(TestCase):
         self.assertTrue(purchase_object_mock.save.called)
 
         purchase_model_mock.assert_called_with(purchaser=purchaser_object, item=item_object,
-            merchant=merchant_object, count=purchase_count)
+                merchant=merchant_object, count=purchase_count)
